move execution tracing provider back to common_lib
